requirements.txt
loguru==0.5.3
psycopg2-binary==2.9.2
python-snap7==1.1

aiohttp~=3.8.1
redis~=4.0.2